Add interactive requirements
