+++ src/index.html
@@ -40,6 +40,6 @@
         </div>
       </div>
     </main>
-    <script src="bundle.js" type="module"></script>
+    <!-- <script src="bundle.js" type="module"></script> -->
   </body>
 </html>
