+++ src/modules/completed.js
@@ -1,0 +1,9 @@
+const markAsCompleted = (task) => {
+  task.completed = true;
+};
+
+const markAsIncomplete = (task) => {
+  task.completed = false;
+};
+
+export { markAsCompleted, markAsIncomplete };
